requirements.txt
fastapi
uvicorn
psycopg[binary,pool]
python-dotenv
pydantic[email]
bcrypt
python-jose[cryptography]
python-multipart
scikit-learn
pandas
numpy
openai
joblib
psycopg-pool